pyproject.toml
[project]
name = "rag-system"
version = "0.1.0"
description = "基于LangGraph的RAG系统，支持文档上传、向量检索和智能问答"
authors = [
    {name = "RAG Team", email = "team@rag.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.10",
    "langgraph>=0.0.20",
    "langsmith>=0.0.69",
    "faiss-cpu>=1.7.4",
    "pymilvus>=2.3.0",
    "chromadb>=0.4.18",
    "pypdf>=3.17.0",
    "python-docx>=1.1.0",
    "markdown>=3.5.1",
    "sentence-transformers>=2.2.2",
    "numpy>=1.24.0",
    "pandas>=2.1.0",
    "loguru>=0.7.2",
    "python-dotenv>=1.0.0",
    "redis>=5.0.0",
    "celery>=5.3.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0"
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.7.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
